Add existing file

+++ NewsNow/manager/views.py
@@ -13,13 +13,41 @@ from subcat.models import SubCat
 from trending.models import Trending
 
 def manager_list(request):
+    # login check start
+    if not request.user.is_authenticated:
+        return redirect('my_login')
+    # login check end
+    # check for permission
+    perm = 0
+    for i in request.user.groups.all():
+
+        if i.name == "masteruser":
+            perm = 1
+    if perm == 0:
+        error = "Access Denied"
+        return render(request, 'back/error.html', {'error': error})
+    # end check for permission
+
 
     manager=Manager.objects.all()
 
     return render(request,'back/manager/manager_list.html',{'manager':manager})
 
 def manager_del(request,pk):
+    # login check start
+    if not request.user.is_authenticated:
+        return redirect('my_login')
+    # login check end
+    # check for permission
+    perm = 0
+    for i in request.user.groups.all():
 
+        if i.name == "masteruser":
+            perm = 1
+    if perm == 0:
+        error = "Access Denied"
+        return render(request, 'back/error.html', {'error': error})
+    # end check for permission
 
     manager=Manager.objects.get(pk=pk)
     b=User.objects.filter(username=manager.utxt)
@@ -32,13 +60,39 @@ def manager_del(request,pk):
     return redirect('manager_list')
 
 def manager_group(request):
+    # login check start
+    if not request.user.is_authenticated:
+        return redirect('my_login')
+    # login check end
 
-    group=Group.objects.all()
+    perm=0
+    for i in request.user.groups.all():
+
+        if i.name =="masteruser":
+            perm=1
+    if perm == 0:
+        error = "Access Denied"
+        return render(request, 'back/error.html', {'error': error})
+
+    group=Group.objects.all().exclude(name='masteruser')
 
     return render(request,'back/manager/manager_group.html',{'group':group})
 
 def manager_group_add(request):
+    # login check start
+    if not request.user.is_authenticated:
+        return redirect('my_login')
+    # login check end
+    #check for permission
+    perm = 0
+    for i in request.user.groups.all():
 
+        if i.name == "masteruser":
+            perm = 1
+    if perm == 0:
+        error = "Access Denied"
+        return render(request, 'back/error.html', {'error': error})
+    #end check for permission
     if request.method=="POST":
         name=request.POST.get('name')
         if  name!="":
@@ -52,6 +106,10 @@ def manager_group_add(request):
     return redirect('manager_group')
 
 def manager_group_del(request,name):
+    # login check start
+    if not request.user.is_authenticated:
+        return redirect('my_login')
+    # login check end
 
     b=Group.objects.filter(name=name)
     b.delete()
@@ -59,6 +117,21 @@ def manager_group_del(request,name):
     return redirect('manager_group')
 
 def user_groups(request,pk):
+    # login check start
+    if not request.user.is_authenticated:
+        return redirect('my_login')
+    # login check end
+
+    # check for permission
+    perm = 0
+    for i in request.user.groups.all():
+
+        if i.name == "masteruser":
+            perm = 1
+    if perm == 0:
+        error = "Access Denied"
+        return render(request, 'back/error.html', {'error': error})
+    # end check for permission
 
     manager=Manager.objects.get(pk=pk)
     print(manager.utxt)
@@ -71,10 +144,14 @@ def user_groups(request,pk):
 
     group=Group.objects.all()
 
-    return render(request,'back/manager/user_groups.html',{'ugroup':ugroup,'group':group})
+    return render(request,'back/manager/user_groups.html',{'ugroup':ugroup,'group':group,'pk':pk})
 
 
 def add_user_to_groups(request,pk):
+    # login check start
+    if not request.user.is_authenticated:
+        return redirect('my_login')
+    # login check end
 
     if request.method=="POST":
 
@@ -84,6 +161,19 @@ def add_user_to_groups(request,pk):
         user=User.objects.get(username=manager.utxt)
         user.groups.add(group)
 
-    return redirect('user_groups',{pk:pk})
+    return redirect('user_groups',pk=pk)
 
     #return render(request,'back/manager/user_groups.html')
+
+def del_user_from_groups(request,pk,name):
+    # login check start
+    if not request.user.is_authenticated:
+        return redirect('my_login')
+    # login check end
+
+    group=Group.objects.get(name=name)#getting the group with the name
+    manager=Manager.objects.get(pk=pk)#get manager corresponding with pk
+    user=User.objects.get(username=manager.utxt)#get user with username=manager.utxt
+    user.groups.remove(group) #remove group
+
+    return redirect('user_groups',pk=pk)
